Split DB into a separate package
